requirements.txt
pytest==7.4.4
selene>=2.0.0rc2
allure-pytest==2.13.2
selenium==4.16.0

